refactor resolvers to use the new services
