+++ subschemas/substrate-api-proxy/src/resolvers/Query/parachains.ts
@@ -1,14 +1,10 @@
-import { ApiPromise } from '@polkadot/api';
 import type { Option, StorageKey } from '@polkadot/types';
-import type { ParaId, BlockNumber } from '@polkadot/types/interfaces';
+import type { ParaId } from '@polkadot/types/interfaces';
 import type { PolkadotRuntimeParachainsParasParaLifecycle } from '@polkadot/types/lookup';
 import type { Context } from '../../types';
-import type {
-  ParachainsInfo,
-  LeasePeriod,
-} from '../../generated/resolvers-types';
-import { BN_ZERO, formatNumber, BN_ONE, BN_HUNDRED } from '@polkadot/util';
-import { getBlockTime } from '../../utils/blockTime';
+import type { ParachainsInfo } from '../../generated/resolvers-types';
+
+import { getLeasePeriod } from '../../services/parachainsService';
 
 type ParaIdEntries = [
   StorageKey<[ParaId]>,
@@ -58,28 +54,4 @@ function extractUpcomingParaIds(entries: ParaIdEntries = []): ParaId[] {
     )
     .filter((paraId): paraId is ParaId => !!paraId)
     .sort((a, b) => a.cmp(b));
-}
-
-async function getLeasePeriod(api: ApiPromise): Promise<LeasePeriod> {
-  const bestNumber = await api.derive.chain.bestNumber();
-  const leasePeriodLength = api.consts.slots.leasePeriod as BlockNumber;
-  const { formattedTime: totalPeriod } = getBlockTime(api, leasePeriodLength);
-  const startNumber = bestNumber.sub(
-    (api.consts.slots.leaseOffset as BlockNumber) || BN_ZERO,
-  );
-  const currentPeriod = startNumber.div(leasePeriodLength);
-  const progress = startNumber.mod(leasePeriodLength);
-  const progressPercent = progress
-    .mul(BN_HUNDRED)
-    .div(leasePeriodLength ?? BN_ONE)
-    .toNumber();
-  const periodRemainder = leasePeriodLength.sub(progress);
-  const { formattedTime: remainder } = getBlockTime(api, periodRemainder);
-
-  return {
-    currentLease: formatNumber(currentPeriod),
-    totalPeriod,
-    progressPercent,
-    remainder,
-  };
 }
